requirements.txt
absl-py==0.15.0
altgraph==0.17.2
anyio==3.5.0
appdirs==1.4.4
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
audioread==2.1.9
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
cryptography==36.0.2
decorator==4.4.2
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.3.0
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h11==0.12.0
h2==4.1.0
h5py==2.10.0
hpack==4.0.0
httpcore==0.13.7
httpx==0.19.0
hyperframe==6.0.1
idna==3.3
imageio==2.18.0
imageio-ffmpeg==0.4.7
importlib-metadata==4.11.3
joblib==1.1.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
librosa==0.8.0
llvmlite==0.36.0
Markdown==3.3.6
moviepy==1.0.3
norbert==0.2.1
numba==0.53.1
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.4.2
pefile==2021.9.3
Pillow==9.1.0
pooch==1.6.0
proglog==0.1.9
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rfc3986==1.5.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
selenium==4.1.3
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
sox==1.4.1
spleeter==2.3.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==3.1.0
torch==1.8.2+cu111
torchaudio==0.8.2
torchvision==0.9.2+cu111
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
typeguard==2.13.3
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.9
validators==0.18.2
webrtcvad==2.0.10
Werkzeug==2.1.2
wincertstore==0.2
wrapt==1.12.1
wsproto==1.1.0
you-get==0.4.1555
zipp==3.8.0
